requirements.txt
# MENA News Collector & Ranking Service - Python Dependencies

# RSS feed parsing
feedparser>=6.0.10

# Google Cloud Firestore
google-cloud-firestore>=2.16.0

# Scheduling (for scheduler.py)
schedule>=1.2.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# OpenAI (for enhanced summaries and weekly reports)
openai>=1.0.0

# Telegram Bot (for inline keyboards and callbacks)
python-telegram-bot>=20.0
